+++ CHANGELOG.md
@@ -30,6 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Snowflake native data-backend implementation
 - Add redis cache inside `db.metadata` for quick multi-process loading
 - Add redis plugin
+- Add pydantic schema support
 
 #### Bug Fixes